* Added DI/IoC * Fixed bug in input with non-newline single char delimiter
