+++ TO-R365-challenge-calculator/TO-R365-challenge-calculator/TO-R365-challenge-calculator.csproj
@@ -7,6 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Unity.Container" Version="5.11.11" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\TOR365Calculator\TOR365Calculator.csproj" />
   </ItemGroup>
