+++ TO-R365-challenge-calculator/TOR365Calculator/IR365Calculator.cs
@@ -5,10 +5,12 @@
         string[] Delimiters {
             get;
             set;
-        } 
+        }
+
+        public void ResetDelimiters(string[] delimiters);
 
         public void ParseInput(string customDelimiter, string input);
 
-        public string Add();
+        public string Calculate();
     }
 }
